[Kysely] Migration of condition result types off sequelize
